+++ docs/source/conf.py
@@ -403,4 +403,10 @@ texinfo_documents = [
 
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/': None}
+intersphinx_mapping = {
+    'python': ('https://docs.python.org/3/', None),
+    'numpy': ('https://docs.scipy.org/doc/numpy/', None),
+    'scipy': ('https://docs.scipy.org/doc/scipy/reference/', None),
+    'xlmhg': ('https://xl-mhg.readthedocs.io/en/latest/', None),
+    'genometools': ('https://genometools.readthedocs.io/en/latest/', None),
+}