+++ gopca/__init__.py
@@ -11,7 +11,7 @@ from .signature_matrix import GOPCASignatureMatrix
 from .run import GOPCARun
 from .go_pca import GOPCA
 
-version = text(pkg_resources.require('gopca')[0].version)
+__version__ = text(pkg_resources.require('gopca')[0].version)
 
 __all__ = ['GOPCAParams', 'GOPCAConfig', 'GOPCA', 'GOPCARun',
            'GOPCASignatureMatrix', 'GOPCASignature']